Build/compiler fix for incompatible pointer types gcc version 14.1.1 …

+++ src/ossimPlanet/unzip.c
@@ -35,6 +35,7 @@ woven in by Terry Thorsen 1/2003.
  */
 
 
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -147,7 +148,8 @@ typedef struct
     int encrypted;
 #    ifndef NOUNCRYPT
     unsigned long keys[3];     /* keys defining the pseudo-random sequence */
-    const unsigned long* pcrc_32_tab;
+    // const unsigned long* pcrc_32_tab;
+    const uint32_t* pcrc_32_tab;
 #    endif
 } unz_s;
